refactor: moved generate_stale_sch_logs to dao
