+++ tdp/cli/commands/status/generate_stales.py
@@ -51,7 +51,7 @@ def generate_stales(
     check_services_cleanliness(cluster_variables)
 
     with Dao(db_engine) as dao:
-        stale_status_logs = dao.get_cluster_status().generate_stale_sch_logs(
+        stale_status_logs = dao.generate_stale_sch_logs(
             cluster_variables=cluster_variables, collections=collections
         )
